\name{lcy.attr.check}
\alias{lcy.attr.check}
\title{
    check existance of attributes of vertex and/or edge
}
\description{
    used to check whether vertex and/or edge attribute exists in the graph
}
\usage{
lcy.attr.check(g, v.attr, e.attr)
}
\arguments{
  \item{g}{ an igraph object
}
  \item{v.attr}{ a vector of characters of vertex attributes to be checked the existance
}
  \item{e.attr}{ a vector of characters of edge attributes to be checked the existance
}
}
\keyword{ attribute }
\keyword{ check }
\keyword{ lcy }
