\name{lcy.graph.plot}
\alias{lcy.graph.plot}
\title{
    plot igraph
}
    
\description{
    customize igraph plot, see igraph documentation. all these parameters can be set as graph attributes without prefix.
}
\usage{
lcy.graph.plot(g, layout = layout.fruchterman.reingold, vertex.size = 2, vertex.label = NA, vertex.label.degree = pi/4, vertex.label.font = 1, vertex.label.cex = 0.3, vertex.label.color = "#0000FF", vertex.frame.color = "black", vertex.label.dist = 0.1, edge.arrow.size = 0.2, edge.width = 0.2, edge.label = NA, edge.label.font = 1, edge.label.cex = 0.3, main = "", arg.legend = NULL)
}
\arguments{
  \item{g}{
}
  \item{layout}{
}
  \item{vertex.size}{
}
  \item{vertex.label}{
}
  \item{vertex.label.degree}{
}
  \item{vertex.label.font}{
}
  \item{vertex.label.cex}{
}
  \item{vertex.label.color}{
}
  \item{vertex.frame.color}{ vertex.shape is also important.
}
  \item{vertex.label.dist}{
}
  \item{edge.arrow.size}{
}
  \item{edge.width}{
}
  \item{edge.label}{
}
  \item{edge.label.font}{
}
  \item{edge.label.cex}{
}
  \item{main}{
}
  \item{arg.legend}{
}
}
\keyword{ plot }
